Implemented TurboSQL for increased speed!
Previous releases created a "dummy" table in the SQL database.  This
release utilizes the TurboSQL switch on the file driver, which does not
create a table in the SQL database and improves speed and performance!

Other improvements include the login procedure will now exit cleanly
when clicking "cancel", and the TestConnection method now tests to make
sure the database exists (previously only checked that you could login
to the server).